#6 - refactor: 로그아웃 모달 기능 분리
